refactor: 임시적 초기 분리 작업 (#86)
