+++ src/main/java/com/codenear/butterfly/auth/application/AuthService.java
@@ -6,6 +6,7 @@ import com.codenear.butterfly.auth.application.jwt.JwtService;
 import com.codenear.butterfly.auth.domain.dto.AuthRequestDTO;
 import com.codenear.butterfly.auth.domain.dto.CustomUserDetails;
 import com.codenear.butterfly.auth.exception.AuthException;
+import com.codenear.butterfly.auth.exception.message.MessageUtil;
 import com.codenear.butterfly.global.exception.ErrorCode;
 import com.codenear.butterfly.member.domain.Grade;
 import com.codenear.butterfly.member.domain.Member;
@@ -35,8 +36,8 @@ public class AuthService {
     public void handleRegistration(AuthRequestDTO requestDTO) {
         Map<Platform, Runnable> platformActions = Map.of(
                 Platform.CODENEAR, () -> emailRegisterService.emailRegister(requestDTO),
-                Platform.KAKAO, () -> registerOrLogin(requestDTO),
-                Platform.GOOGLE, () -> registerOrLogin(requestDTO)
+//                Platform.KAKAO, () -> registerOrLogin(requestDTO),
+//                Platform.GOOGLE, () -> registerOrLogin(requestDTO)
         );
 
         Optional.ofNullable(platformActions.get(requestDTO.getPlatform()))
@@ -48,12 +49,12 @@ public class AuthService {
         try {
             Map<Platform, Runnable> loginActions = Map.of(
                     Platform.CODENEAR, () -> emailLoginAndIssueJwt(requestDTO, response),
-                    Platform.KAKAO, () -> socialLoginAndIssueJwt(requestDTO, response),
-                    Platform.GOOGLE, () -> socialLoginAndIssueJwt(requestDTO, response)
+//                    Platform.KAKAO, () -> socialLoginAndIssueJwt(requestDTO, response),
+//                    Platform.GOOGLE, () -> socialLoginAndIssueJwt(requestDTO, response)
             );
 
             Optional.ofNullable(loginActions.get(requestDTO.getPlatform()))
-                    .orElseThrow(() -> new AuthException(ErrorCode.INVALID_PLATFORM, requestDTO.getPlatform()))
+                    .orElseThrow(() -> new IllegalArgumentException("제공하지 않는 플랫폼입니다."))
                     .run();
 
         } catch (BadCredentialsException e) {
@@ -66,31 +67,7 @@ public class AuthService {
         issueJwtTokens(loginUser.getEmail(), requestDTO.getPlatform().name(), response);
     }
 
-    private void socialLoginAndIssueJwt(AuthRequestDTO requestDTO, HttpServletResponse response) {
-        registerOrLogin(requestDTO);
-        issueJwtTokens(requestDTO.getEmail(), requestDTO.getPlatform().name(), response);
-    }
-
     private void issueJwtTokens(String email, String platform, HttpServletResponse response) {
         jwtService.processTokens(email, platform, response);
-    }
-
-    public void registerOrLogin(AuthRequestDTO requestDTO) {
-        Optional<Member> optMember = memberRepository.findByEmailAndPlatform(requestDTO.getEmail(), requestDTO.getPlatform());
-
-        if (optMember.isEmpty()) {
-            Member registerMember = register(requestDTO);
-            memberRepository.save(registerMember);
-        }
-    }
-
-    private Member register(AuthRequestDTO requestDTO) {
-        return Member.builder()
-                .email(requestDTO.getEmail())
-                .nickname(requestDTO.getNickname())
-                .point(0)
-                .grade(Grade.EGG)
-                .platform(requestDTO.getPlatform())
-                .build();
     }
 }
